refactor : Update convertToEstimateResponse task

+++ src/main/java/com/example/daon/receipts/dto/request/ReceiptRequest.java
@@ -9,6 +9,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -33,7 +34,7 @@ public class ReceiptRequest {
 
     private Integer quantity; // 사용 품목 수량
 
-    private int totalPrice; // 품목 총 가격
+    private BigDecimal totalPrice; // 품목 총 가격
 
     private String description; // 전표 설명 (적요)
