+++ src/main/java/com/example/daon/receipts/dto/response/ReceiptResponse.java
@@ -32,7 +32,7 @@ public class ReceiptResponse {
 
     private Integer quantity; // 사용 품목 수량
 
-    private int totalPrice; // 품목 총 가격
+    private BigDecimal totalPrice; // 품목 총 가격
 
     private String description; // 전표 설명 (적요)
