+++ src/main/java/com/example/daon/receipts/model/ReceiptEntity.java
@@ -11,6 +11,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.GenericGenerator;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -51,7 +52,7 @@ public class ReceiptEntity {
     private Integer quantity; // 사용 품목 수량
 
     @Column(name = "total_price", nullable = false)
-    private int totalPrice; // 품목 총 가격
+    private BigDecimal totalPrice; // 품목 총 가격
 
     @Column(name = "description", columnDefinition = "TEXT")
     private String description; // 전표 설명 (적요)
